+++ webapp/channels/src/components/advanced_text_editor/send_button/scheduled_post_custom_time_modal/dmUserTimezone.scss
@@ -1,3 +1,6 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
 .DMUserTimezone {
     margin-top: 8px;
     color: rgba(var(--center-channel-color-rgb), 0.75);
